requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
PyAudio==0.2.14
regex==2024.11.6
requests==2.32.3
setuptools==78.1.0
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.1
urllib3==2.3.0
